requirements.txt
fastapi
uvicorn[standard]
google-genai
snowflake-connector-python
python-dotenv
